console/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
google-cloud-bigquery>=3.20.0
google-cloud-pubsub>=2.21.0
google-cloud-firestore>=2.16.0
google-auth>=2.29.0
python-multipart>=0.0.9
